Extract analyzer and code fix base classes for block-level documentation analysis
