+++ StyleCop.Analyzers/StyleCop.Analyzers.Test/DocumentationRules/SA1653UnitTests.cs
@@ -322,7 +322,7 @@ public class ClassName
 
         protected override CodeFixProvider GetCSharpCodeFixProvider()
         {
-            return new SA1653CodeFixProvider();
+            return new BlockLevelDocumentationCodeFixProvider();
         }
     }
 }
